--- constants/theme.ts
@@ -1,31 +1,0 @@
-const system = {
-  warning: "#FFBB2E",
-  error: "#CC0000",
-  success: "#28C840",
-  successLight: "#D4F4D9",
-  focus: "#1460cc",
-  disabled: "#A9A9A9",
-};
-
-export const theme = {
-  light: {
-    text: "black",
-    black: "#000000",
-    white: "#FFFFFF",
-    primary: "#764CD2",
-    secondary: "#FF6048",
-    contrast: "#FFFFFF",
-    background: "#F7F9FC",
-    system,
-  },
-  dark: {
-    text: "black",
-    black: "#000000",
-    white: "#FFFFFF",
-    primary: "#764CD2",
-    secondary: "#FF6048",
-    contrast: "#FFFFFF",
-    background: "#F7F9FC",
-    system,
-  },
-};
